cleaned up home page
