--- client/src/pages/SearchBooks.js
@@ -1,211 +1,0 @@
-import React, { useState } from 'react';
-import { Jumbotron, Container, Col, Form, Button, Card, CardColumns } from 'react-bootstrap';
-
-import Auth from '../utils/auth';
-import { useMutation, useQuery } from '@apollo/client';
-import { SAVE_BOOK, ADD_TODO } from '../utils/mutations';
-import { searchGoogleBooks } from '../utils/API';
-import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
-import { QUERY_ME } from '../utils/queries';
-import ToDo from '../components/ToDo';
-
-const SearchBooks = () => {
-  const [saveBook] = useMutation(SAVE_BOOK);
-  // create state for holding returned google api data
-  const [searchedBooks, setSearchedBooks] = useState([]);
-  // create state for holding our search field data
-  const [searchInput, setSearchInput] = useState('');
-
-
-  // create state to hold saved bookId values
-  const [savedBookIds, setSavedBookIds] = useState(getSavedBookIds());
-
-  // create method to search for books and set state on form submit
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-
-    if (!searchInput) {
-      return false;
-    }
-
-    try {
-      const response = await searchGoogleBooks(searchInput);
-
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
-
-      const { items } = await response.json();
-
-      const bookData = items.map((book) => ({
-        bookId: book.id,
-        authors: book.volumeInfo.authors || ['No author to display'],
-        title: book.volumeInfo.title,
-        description: book.volumeInfo.description,
-        image: book.volumeInfo.imageLinks?.thumbnail || '',
-      }));
-
-      setSearchedBooks(bookData);
-      setSearchInput('');
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  // create function to handle saving a book to our database
-  const handleSaveBook = async (bookId) => {
-    // find the book in `searchedBooks` state by the matching id
-    const bookToSave = searchedBooks.find((book) => book.bookId === bookId);
-
-    // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-    if (!token) {
-      return false;
-    }
-
-    try {
-      await saveBook({ variables: bookToSave })
-      // if book successfully saves to user's account, save book id to state
-      setSavedBookIds([...savedBookIds, bookToSave.bookId]);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  // create state for holding our search field data
-  const [toDoInput, setToDoInput] = useState('');
-  //useQuery for querying users to-dos
-  const {loading, data} = useQuery(QUERY_ME);
-  //set toDos = returned data from me query in toDo array of user
-  const toDos = data?.me.toDo || [];
-  //create const of addToDo that holds ADD_TODO mutation
-  const [addToDo] = useMutation(ADD_TODO);
-
-
-
-      // create function to handle saving a todo to our database
-  const handleAddToDo = async (event) => {
-    // find the book in `searchedBooks` state by the matching id
-    //const bookToSave = searchedBooks.find((book) => book.bookId === bookId);\
-    event.preventDefault();
-    console.log(toDoInput);
-
-    // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-    //if user not logged in, return out
-    if (!token) {
-      return <h3>Please Login to Add To-Dos</h3>;
-    }
-
-    try {
-      //add todo to DB by passing toDoInput state variable as text attribute
-      await addToDo({ variables: {text: toDoInput} })
-    } catch (err) {
-      console.log('we got an error bud');
-      console.error(err);
-    }
-    //clear ToDo Input value
-    setToDoInput('')
-  };
-
-  return (
-    <>
-      <Jumbotron fluid className='text-light bg-dark'>
-        <Container>
-          <h1>Search for Books!</h1>
-          <Form onSubmit={handleFormSubmit}>
-            <Form.Row>
-              <Col xs={12} md={8}>
-                <Form.Control
-                  name='searchInput'
-                  value={searchInput}
-                  onChange={(e) => setSearchInput(e.target.value)}
-                  type='text'
-                  size='lg'
-                  placeholder='Search for a book'
-                />
-              </Col>
-              <Col xs={12} md={4}>
-                <Button type='submit' variant='success' size='lg'>
-                  Submit Search
-                </Button>
-              </Col>
-            </Form.Row>
-          </Form>
-        </Container>
-        <Container>
-          <h1>Add Your To-Dos!</h1>
-          <Form onSubmit={handleAddToDo}>
-            <Form.Row>
-              <Col xs={12} md={8}>
-                <Form.Control
-                  name='toDoInput'
-                  value={toDoInput}
-                  onChange={(e) => setToDoInput(e.target.value)}
-                  type='text'
-                  size='lg'
-                  placeholder='What do you need to do?'
-                />
-              </Col>
-              <Col xs={12} md={4}>
-                <Button type='submit' variant='success' size='lg'>
-                  Add
-                </Button>
-              </Col>
-            </Form.Row>
-          </Form>
-        </Container>
-      </Jumbotron>
-
-      <Container>
-
-      {loading ? (
-        <div>Loading...</div>
-      ) : (
-        <ToDo
-        toDos={toDos}></ToDo>
-      )}
-
-
-
-
-
-        <h2>
-          {searchedBooks.length
-            ? `Viewing ${searchedBooks.length} results:`
-            : 'Search for a book to begin'}
-        </h2>
-        <CardColumns>
-          {searchedBooks.map((book) => {
-            return (
-              <Card key={book.bookId} border='dark'>
-                {book.image ? (
-                  <Card.Img src={book.image} alt={`The cover for ${book.title}`} variant='top' />
-                ) : null}
-                <Card.Body>
-                  <Card.Title>{book.title}</Card.Title>
-                  <p className='small'>Authors: {book.authors}</p>
-                  <Card.Text>{book.description}</Card.Text>
-                  {Auth.loggedIn() && (
-                    <Button
-                      disabled={savedBookIds?.some((savedBookId) => savedBookId === book.bookId)}
-                      className='btn-block btn-info'
-                      onClick={() => handleSaveBook(book.bookId)}>
-                      {savedBookIds?.some((savedBookId) => savedBookId === book.bookId)
-                        ? 'This book has already been saved!'
-                        : 'Save this Book!'}
-                    </Button>
-                  )}
-                </Card.Body>
-              </Card>
-            );
-          })}
-        </CardColumns>
-      </Container>
-    </>
-  );
-};
-
-export default SearchBooks;
